requirements.txt
# Python dependencies for AI Research Agent

# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Web Scraping and Content Extraction
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
trafilatura==1.7.0
readability-lxml==0.8.1

# Search APIs
google-search-results==2.4.2  # SerpAPI
requests-oauthlib==1.3.1

# Natural Language Processing
transformers==4.34.0
torch==2.0.1
tokenizers==0.13.3

# Optional AI APIs (uncomment as needed)
# openai==0.28.1
# anthropic==0.7.7

# Text Processing
nltk==3.8.1
spacy==3.7.1
textstat==0.7.3

# PDF and Document Processing
PyPDF2==3.0.1
pymupdf==1.23.8  # fitz
reportlab==4.0.4
markdown==3.5.1
python-docx==0.8.11

# Export and File Handling
weasyprint==60.1
markdown-pdf==0.1.1
jinja2==3.1.2

# Database and Caching (optional)
redis==5.0.1
sqlite3  # Built-in with Python

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
urllib3==2.0.5
chardet==5.2.0
lxml==4.9.3

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3

# Environment and Configuration
python-decouple==3.8
configparser==6.0.0

# Logging and Monitoring
python-json-logger==2.0.7

# Security
cryptography==41.0.7
bcrypt==4.0.1

# Optional Vector Database (for RAG)
# faiss-cpu==1.7.4
# chromadb==0.4.13

# Optional Academic APIs
# scholarly==1.7.11
# crossref-commons==0.10.0

# Citation Processing
bibtexparser==1.4.0
citeproc-py==0.6.0

# Image Processing (if needed)
Pillow==10.0.1

# Async Processing
celery==5.3.3
redis==5.0.1  # For Celery broker

# Data Validation
cerberus==1.3.5
marshmallow==3.20.1

# HTTP Client Enhancements
httpx==0.25.0
aiohttp==3.8.6

# Text Similarity and Analysis
sentence-transformers==2.2.2
scikit-learn==1.3.1

# Rate Limiting
Flask-Limiter==3.5.0

# CORS and Security Headers
flask-talisman==1.1.0

# Environment Detection
python-magic==0.4.27

# Progress Bars and UI
tqdm==4.66.1

# Time and Date Handling
arrow==1.3.0
pendulum==2.1.2

# Validation and Sanitization
bleach==6.1.0
html5lib==1.1

# Memory and Performance
psutil==5.9.6
memory-profiler==0.61.0

# Error Tracking (optional)
# sentry-sdk[flask]==1.35.0

# API Documentation (optional)
# flask-restx==1.2.0
# flasgger==0.9.7.1